myproject/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.140
botocore==1.34.140
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-storages==1.14.3
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
groupdocs-conversion-cloud==24.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
Jinja2==3.1.4
jmespath==1.0.1
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.3.7
konlpy==0.6.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.12
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.84
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
moto==5.0.11
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
openai==1.35.10
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pdfrw==0.4
pillow==10.4.0
pinecone==4.0.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyOpenSSL==24.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
responses==0.25.3
s3transfer==0.10.2
safetensors==0.4.3
service-identity==24.1.0
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
sympy==1.13.0
tenacity==8.5.0
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.3
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
xmltodict==0.13.0
yarl==1.9.4
zope.interface==6.4.post2
cryptography==42.0.8